+++ sqflite/lib/src/batch.dart
@@ -1,18 +1,35 @@
 import 'package:sqflite/sqlite_api.dart';
 import 'package:sqflite/src/constant.dart';
 import 'package:sqflite/src/database.dart';
+import 'package:sqflite/src/sqflite_impl.dart';
 import 'package:sqflite/src/sql_builder.dart';
 import 'package:sqflite/src/transaction.dart';
 
+/// Batch implementation
 abstract class SqfliteBatch implements Batch {
+  /// List of operations
   final List<Map<String, dynamic>> operations = <Map<String, dynamic>>[];
 
-  void _add(String method, String sql, List<dynamic> arguments) {
-    operations.add(<String, dynamic>{
+  Map<String, dynamic> _getOperationMap(
+      String method, String sql, List<dynamic> arguments) {
+    return <String, dynamic>{
       paramMethod: method,
       paramSql: sql,
       paramSqlArguments: arguments
-    });
+    };
+  }
+
+  void _add(String method, String sql, List<dynamic> arguments) {
+    operations.add(_getOperationMap(method, sql, arguments));
+  }
+
+  void _addExecute(
+      String method, String sql, List<dynamic> arguments, bool inTransaction) {
+    final Map<String, dynamic> map = _getOperationMap(method, sql, arguments);
+    if (inTransaction != null) {
+      map[paramInTransaction] = inTransaction;
+    }
+    operations.add(map);
   }
 
   @override
@@ -88,13 +105,18 @@ abstract class SqfliteBatch implements Batch {
 
   @override
   void execute(String sql, [List<dynamic> arguments]) {
-    _add(methodExecute, sql, arguments);
+    // Check for begin/end transaction
+    final bool inTransaction = getSqlInTransactionArgument(sql);
+    _addExecute(methodExecute, sql, arguments, inTransaction);
   }
 }
 
+/// Batch on a given database
 class SqfliteDatabaseBatch extends SqfliteBatch {
+  /// Create a batch in a database
   SqfliteDatabaseBatch(this.database);
 
+  /// Our database
   final SqfliteDatabase database;
 
   @override
@@ -109,9 +131,12 @@ class SqfliteDatabaseBatch extends SqfliteBatch {
   }
 }
 
+/// Batch on a given transaction
 class SqfliteTransactionBatch extends SqfliteBatch {
+  /// Create a batch in a transaction
   SqfliteTransactionBatch(this.transaction);
 
+  /// Our transaction
   final SqfliteTransaction transaction;
 
   @override