+++ sqflite/lib/src/database_mixin.dart
@@ -3,17 +3,21 @@ import 'dart:async';
 import 'package:sqflite/sqlite_api.dart';
 import 'package:sqflite/src/batch.dart';
 import 'package:sqflite/src/collection_utils.dart';
-import 'package:sqflite/src/constant.dart' hide lockWarningDuration;
+import 'package:sqflite/src/constant.dart';
 import 'package:sqflite/src/database.dart';
 import 'package:sqflite/src/exception.dart';
 import 'package:sqflite/src/factory.dart';
 import 'package:sqflite/src/sql_builder.dart';
 import 'package:sqflite/src/transaction.dart';
+import 'package:sqflite/src/utils.dart';
 import 'package:sqflite/utils/utils.dart';
+import 'package:sqflite/src/utils.dart' as utils;
 import 'package:synchronized/synchronized.dart';
 
+/// Base database implementation
 class SqfliteDatabaseBase
     with SqfliteDatabaseMixin, SqfliteDatabaseExecutorMixin {
+  /// ctor
   SqfliteDatabaseBase(SqfliteDatabaseOpenHelper openHelper, String path,
       {OpenDatabaseOptions options}) {
     this.openHelper = openHelper;
@@ -21,6 +25,7 @@ class SqfliteDatabaseBase
   }
 }
 
+/// Common database/transaction implementation
 mixin SqfliteDatabaseExecutorMixin implements SqfliteDatabaseExecutor {
   @override
   SqfliteTransaction get txn;
@@ -182,7 +187,7 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   Future<T> safeInvokeMethod<T>(String method, [dynamic arguments]) =>
       factory.wrapDatabaseException(() => invokeMethod(method, arguments));
 
-  // save the open helper for proper closing
+  /// Keep our open helper for proper closing.
   SqfliteDatabaseOpenHelper openHelper;
   @override
   OpenDatabaseOptions options;
@@ -190,9 +195,6 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   SqfliteDatabaseFactory get factory => openHelper.factory;
 
   bool get readOnly => openHelper?.options?.readOnly == true;
-
-  // Set when parsing BEGIN and COMMIT/ROLLBACK
-  bool inTransaction = false;
 
   @override
   SqfliteDatabase get db => this;
@@ -207,27 +209,42 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   @override
   String path;
 
-  // only set during inTransaction to allow transaction during open
+  /// Transaction reference count.
+  ///
+  /// Only set during inTransaction to allow transaction during open.
   int transactionRefCount = 0;
 
-  // Not null during opening
-  // default transaction used during opening
+  /// Special transaction created during open.
+  ///
+  /// Only not null during opening.
   SqfliteTransaction openTransaction;
 
   @override
   SqfliteTransaction get txn => openTransaction;
 
-  // non-reentrant lock
+  /// Non-reentrant lock.
   final Lock rawLock = Lock();
 
   // Its internal id
   @override
   int id;
 
+  /// Set when parsing BEGIN and COMMIT/ROLLBACK
+  bool inTransaction = false;
+
   static Map<String, dynamic> getBaseDatabaseMethodArguments(int id) {
     final Map<String, dynamic> map = <String, dynamic>{
       paramId: id,
     };
+    return map;
+  }
+
+  static Map<String, dynamic> getBaseDatabaseMethodArgumentsInTransaction(
+      int id, bool inTransaction) {
+    final Map<String, dynamic> map = getBaseDatabaseMethodArguments(id);
+    if (inTransaction != null) {
+      map[paramInTransaction] = inTransaction;
+    }
     return map;
   }
 
@@ -275,8 +292,8 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
       return await action(txn);
     } else {
       // Simple timeout warning if we cannot get the lock after XX seconds
-      final bool handleTimeoutWarning =
-          (lockWarningDuration != null && lockWarningCallback != null);
+      final bool handleTimeoutWarning = (utils.lockWarningDuration != null &&
+          utils.lockWarningCallback != null);
       Completer<dynamic> timeoutCompleter;
       if (handleTimeoutWarning) {
         timeoutCompleter = Completer<dynamic>();
@@ -292,8 +309,9 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
       // Simply warn the developer as this could likely be a deadlock
       if (handleTimeoutWarning) {
         // ignore: unawaited_futures
-        timeoutCompleter.future.timeout(lockWarningDuration, onTimeout: () {
-          lockWarningCallback();
+        timeoutCompleter.future.timeout(utils.lockWarningDuration,
+            onTimeout: () {
+          utils.lockWarningCallback();
         });
       }
       return await operation;
@@ -311,24 +329,29 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   Future<T> txnExecute<T>(SqfliteTransaction txn, String sql,
       [List<dynamic> arguments]) {
     return txnWriteSynchronized<T>(txn, (_) {
-      if (sql != null) {
-        final String lowerSql = sql.trim().toLowerCase();
-        if (lowerSql.startsWith('begin')) {
-          inTransaction = true;
-        } else if (lowerSql.startsWith('commit') ||
-            lowerSql.startsWith('rollback')) {
-          inTransaction = false;
-        }
+      bool inTransactionChange = getSqlInTransactionArgument(sql);
+
+      if (inTransactionChange ?? false) {
+        inTransactionChange = true;
+        inTransaction = true;
+      } else if (inTransactionChange == false) {
+        inTransactionChange = false;
+        inTransaction = false;
       }
-      return invokeExecute<T>(sql, arguments);
+
+      return invokeExecute<T>(sql, arguments,
+          inTransactionChange: inTransactionChange);
     });
   }
 
-  Future<T> invokeExecute<T>(String sql, List<dynamic> arguments) {
+  /// [inTransactionChange] is true when entering a transaction, false when leaving
+  Future<T> invokeExecute<T>(String sql, List<dynamic> arguments,
+      {bool inTransactionChange}) {
     return safeInvokeMethod(
         methodExecute,
-        <String, dynamic>{paramSql: sql, paramSqlArguments: arguments}
-          ..addAll(baseDatabaseMethodArguments));
+        <String, dynamic>{paramSql: sql, paramSqlArguments: arguments}..addAll(
+            getBaseDatabaseMethodArgumentsInTransaction(
+                id, inTransactionChange)));
   }
 
   /// for INSERT sql query
@@ -413,7 +436,7 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
   Future<void> endTransaction(SqfliteTransaction txn) async {
     // never commit transaction in read-only mode
     if (readOnly != true) {
-      if (txn.successfull == true) {
+      if (txn.successful == true) {
         await txnExecute<dynamic>(txn, "COMMIT");
       } else {
         await txnExecute<dynamic>(txn, "ROLLBACK");
@@ -435,7 +458,7 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
     } finally {
       if (--transactionRefCount == 0) {
         final SqfliteTransaction sqfliteTransaction = txn as SqfliteTransaction;
-        sqfliteTransaction.successfull = successfull;
+        sqfliteTransaction.successful = successfull;
         await endTransaction(sqfliteTransaction);
       }
     }
@@ -491,8 +514,9 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
       // create the folder if needed (needed for iOS)
       await factory.createParentDirectory(path);
     }
+    final bool singleInstance = options?.singleInstance ?? false;
     // Single instance?
-    params[paramSingleInstance] = options?.singleInstance != false;
+    params[paramSingleInstance] = singleInstance;
 
     if (options?.password != null){
       params[paramPassword] = options?.password;
@@ -510,20 +534,21 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
     } else if (openResult is Map) {
       final int id = openResult[paramId] as int;
       // Recover means we found an instance in the native world
-      final bool recovered = openResult[paramRecovered] == true;
+      final bool recoveredInTransaction =
+          openResult[paramRecoveredInTransaction] == true;
       // in this case, we are going to rollback any changes in case a transaction
       // was in progress. This catches hot-restart scenario
-      if (recovered) {
+      if (recoveredInTransaction) {
         // Don't do it for read-only
         if (readOnly != true) {
           // We are not yet open so invoke the plugin directly
           try {
             await safeInvokeMethod(
                 methodExecute,
-                <String, dynamic>{paramSql: 'ROLLBACK'}
-                  ..addAll(getBaseDatabaseMethodArguments(id)));
-          } catch (_) {
-            // devPrint('recovered database ROLLBACK failed $e');
+                <String, dynamic>{paramSql: 'ROLLBACK'}..addAll(
+                    getBaseDatabaseMethodArgumentsInTransaction(id, false)));
+          } catch (e) {
+            print('ignore recovered database ROLLBACK error $e');
           }
         }
       }
@@ -553,7 +578,8 @@ mixin SqfliteDatabaseMixin implements SqfliteDatabase {
             await txnWriteSynchronized(txn, (Transaction txn) async {
               // Special trick to cancel any pending transaction
               try {
-                await invokeExecute<dynamic>('ROLLBACK', null);
+                await invokeExecute<dynamic>('ROLLBACK', null,
+                    inTransactionChange: false);
               } catch (_) {
                 // devPrint('rollback error $_');
               }